[GPU] Fix warnings (#16516)
* fix a few warnings

* cast size_t to uint32_t

+++ src/plugins/intel_gpu/src/graph/impls/ocl/arg_max_min.cpp
@@ -66,7 +66,7 @@ public:
         const auto& mode = primitive->mode;
         const auto& sort_type = primitive->sort;
         const auto& values_first = primitive->values_first;
-        const auto& outputs_num = primitive->input_size() == 3 ? 2 : primitive->output_size();
+        const auto& outputs_num = primitive->input_size() == 3 ? 2 : static_cast<uint32_t>(primitive->output_size());
 
         auto argm_params = get_default_params<kernel_selector::arg_max_min_params>(impl_param);
         auto argm_optional_params =
